Added full API for Division and refactored the packages.
