+++ src/main/java/com/keyin/rest/division/DivisionController.java
@@ -1,15 +1,26 @@
-package com.keyin.rest;
+package com.keyin.rest.division;
 
-import com.keyin.domain.Division;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @CrossOrigin
 public class DivisionController {
     @Autowired
     private DivisionService divisionService;
+
+    @GetMapping("/division")
+    public List<Division> getAllDivisions() {
+        return divisionService.getAllDivisions();
+    }
+
+    @GetMapping("/division/{id}")
+    public Division getDivisionById(@PathVariable long id) {
+        return divisionService.getDivisionById(id);
+    }
 
     @PostMapping("/division")
     public Division createDivision(@RequestBody Division division) {
@@ -19,5 +30,10 @@ public class DivisionController {
     @PutMapping("/division/{id}")
     public ResponseEntity<Division> updateDivision(@PathVariable long id, @RequestBody Division division) {
         return ResponseEntity.ok(divisionService.updateDivision(id, division));
+    }
+
+    @DeleteMapping("/division/{id}")
+    public void deleteDivisionById(@PathVariable long id) {
+        divisionService.deleteDivisionById(id);
     }
 }
