+++ src/main/java/com/keyin/rest/division/DivisionService.java
@@ -1,15 +1,29 @@
-package com.keyin.rest;
+package com.keyin.rest.division;
 
-import com.keyin.domain.Division;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
 public class DivisionService {
     @Autowired
     private DivisionRepository divisionRepository;
+
+    public List<Division> getAllDivisions() {
+        return (List<Division>) divisionRepository.findAll();
+    }
+
+    public Division getDivisionById(long id) {
+        Optional<Division> divisionOptional = divisionRepository.findById(id);
+
+        return divisionOptional.orElse(null);
+    }
+
+    public void deleteDivisionById(long id) {
+        divisionRepository.deleteById(id);
+    }
 
     public Division createDivision(Division newDivision) {
         return divisionRepository.save(newDivision);
